Allocation error management for ft_initialize
